feat(shopSim): current simTime in exception, fix neg. downtime (PR #1)
merges PR #1 by ebroda

For an easier debugging of the exception add the current simulation time
to have an approximation “how much” too late the event should be
executed.

Also this now fixes the problem leading to the upper suggestion by
ensuring that the DblStream is not providing a negative number by
drawing a number again until the number is positive - holds for both
streams.

The problem arised on using a (new DblNormal(500, 15)) for both
parameters. Somehow the distribution might also produce negative values
which then lead to an exception in the scheduling (correctly), but this
is never intended. You never want to schedule an event in the past.